requirements.txt
aiohttp==2.3.7
aiotg==0.9.5
aiotools==0.5.0
asyncpg==0.14.0
argh==0.26.2
async-timeout==2.0.0
chardet==3.0.4
dohq-youtrack==0.2.1.44
httplib2==0.10.3
idna==2.6
multidict==3.3.2
pathtools==0.1.2
PyYAML==3.12
tasky==0.8.1
watchdog==0.8.3
yarl==0.17.0
python-env==1.0.0
